+++ src/configs/extra/no-stylistic-rules.ts
@@ -210,6 +210,9 @@ const ALL_STYLISTIC_RULES = {
     'meta-property-ordering': true,
     'test-case-property-ordering': true,
   },
+  'fast-import': {
+    // None
+  },
   'file-progress': {
     // None
   },